+++ projects/new_ui/new_ui.c
@@ -10,7 +10,7 @@ internal_function void winsizecbk(int width, int height)
 {
     set_screen_dimensions(width, height);
 
-    create_orthographic_projection(&main_cam.proj_matrix, 0, width, 0, height, -0.1f, 100.f);
+    create_orthographic_projection(&main_cam.proj_matrix, 0.0f, (float)width, 0.0f,(float)height, -0.1f, 100.f);
     set_projection_matrix(&main_cam.proj_matrix);
 }
